컴포넌트 분리 및 base url 수정
