+++ src/components/PostInput.js
@@ -1,0 +1,8 @@
+export default function PostInput() {
+  return `
+<div class="mb-4 bg-white rounded-lg shadow p-4">
+  <textarea class="w-full p-2 border rounded" placeholder="무슨 생각을 하고 계신가요?"></textarea>
+  <button class="mt-2 bg-blue-600 text-white px-4 py-2 rounded">게시</button>
+</div>
+`;
+}
